refactor: restructure `it` for future method extensions
