+++ fixtures/no-runner/basic-logs.test.ts
@@ -1,7 +1,7 @@
 import { assert } from '../../src/modules/essentials/assert.js';
 import { test } from '../../src/modules/helpers/test.js';
 import { describe } from '../../src/modules/helpers/describe.js';
-import { it } from '../../src/modules/helpers/it.js';
+import { it } from '../../src/modules/helpers/it/core.js';
 
 assert(true, 'Should emit a basic assetion log');
