+++ src/modules/index.ts
@@ -5,7 +5,7 @@ export { assert } from './essentials/assert.js';
 export { strict } from './essentials/strict.js';
 export { test } from './helpers/test.js';
 export { describe } from './helpers/describe.js';
-export { it } from './helpers/it.js';
+export { it } from './helpers/it/core.js';
 export { envFile } from './helpers/env.js';
 export { skip } from './helpers/skip.js';
 export { beforeEach, afterEach } from './helpers/each.js';
