+++ test/e2e/before-and-after-each.test.ts
@@ -1,6 +1,6 @@
 import { test } from '../../src/modules/helpers/test.js';
 import { describe } from '../../src/modules/helpers/describe.js';
-import { it } from '../../src/modules/helpers/it.js';
+import { it } from '../../src/modules/helpers/it/core.js';
 import { poku } from '../../src/modules/essentials/poku.js';
 import { assert } from '../../src/modules/essentials/assert.js';
 import { isProduction } from '../helpers/capture-cli.test.js';
