+++ test/e2e/fail-fast.test.ts
@@ -1,5 +1,5 @@
 import { describe } from '../../src/modules/helpers/describe.js';
-import { it } from '../../src/modules/helpers/it.js';
+import { it } from '../../src/modules/helpers/it/core.js';
 import { assert } from '../../src/modules/essentials/assert.js';
 import { inspectCLI, isProduction } from '../helpers/capture-cli.test.js';
 import { skip } from '../../src/modules/helpers/skip.js';
