+++ test/e2e/runners.test.ts
@@ -1,6 +1,6 @@
 import { execSync } from 'node:child_process';
 import { describe } from '../../src/modules/helpers/describe.js';
-import { it } from '../../src/modules/helpers/it.js';
+import { it } from '../../src/modules/helpers/it/core.js';
 import { assert } from '../../src/modules/essentials/assert.js';
 import { isProduction, inspectCLI } from '../helpers/capture-cli.test.js';
 import { skip } from '../../src/modules/helpers/skip.js';
