+++ test/integration/it/each/each-options-promise.test.ts
@@ -1,5 +1,5 @@
 import { describe } from '../../../../src/modules/helpers/describe.js';
-import { it } from '../../../../src/modules/helpers/it.js';
+import { it } from '../../../../src/modules/helpers/it/core.js';
 import { assert } from '../../../../src/modules/essentials/assert.js';
 import { beforeEach, afterEach } from '../../../../src/modules/helpers/each.js';
