+++ test/integration/it/it.test.ts
@@ -1,6 +1,6 @@
 import { test } from '../../../src/modules/helpers/test.js';
 import { describe } from '../../../src/modules/helpers/describe.js';
-import { it } from '../../../src/modules/helpers/it.js';
+import { it } from '../../../src/modules/helpers/it/core.js';
 
 test('Testing "it" method', () => {
   describe(async () => {
