+++ test/unit/logs.test.ts
@@ -1,6 +1,6 @@
 import process from 'node:process';
 import { describe } from '../../src/modules/helpers/describe.js';
-import { it } from '../../src/modules/helpers/it.js';
+import { it } from '../../src/modules/helpers/it/core.js';
 import { assert } from '../../src/modules/essentials/assert.js';
 import { isQuiet, isDebug, parserOutput } from '../../src/parsers/output.js';
 import { Write } from '../../src/services/write.js';
